requirements.txt
groq
python-dotenv
transformers
torch
regex
